Add files via upload
Description: This C program takes an array of integers as input from the user along with the number of elements in the array and a search key to find if it exists in the array or not.

Algorithm:

Take the input for the number of elements in the array and the array elements from the user. Take the input for the element to be searched. Iterate over the array and compare each element with the search element. If the search element is found, print its index in the array. If the search element is not found, print a message stating that the element is not found. Output: For example, if the input is: How many elements?: 6 Enter array elements: 14 35 61 77 89 45 Enter element to search: 77

Then the output will be: Element found: 3